refactor: update imports and restructure script.js for improved module organization and state management
