+++ src/script.js
